requirements.txt
numpy<=1.26.4
pygame>=2.6.0
pillow>=10.4.0
gymnasium>=0.29.0

# stable-baselines3>=2.3.2
# modelscope
# transformers

# pip install 'ms-swift[llm]' -U